TS-2678: stop inspecting keep alive state to detect truncation
When the server session breaks, is_http_server_eos_truncation()
tries to determine whether the response was truncated. If it was
truncated, then we need to propagate that to the client, otherwise
we can just continue on.

is_http_server_eos_truncation() was examining the current server
keepalive state to determine truncation, which is really not relevant
at all. In this case, the server sent a HTTP/1.1 response without
a Content-Length header and closes the connection. This should not
be classed as a truncation.