TS-2699: Add TSClientProtoStackCreate API
Currently, plugin authors are expected to construct TSClientProtoStack
values by bitshifting them into their correct locations. This is
tedious likely to be error-prone. TSClientProtoStackCreate() accepts
a sequence of TSProtoType values, terminated by a TS_PROTO_NULL.
It combines the values and returns a TSClientProtoStack value that
describes the protocol layering of a client request.