TS-2706: replace the tcp_info plugin config file with options
Managing a configuration file is a lot of pain for a small number
of options. Replace the tcp_info plugin configuration file with
command line options that can be specified in plugin.config. This
also replaces the number is hook mast option with a human-readable
list of hooks where TCP informations whould be logged.