TS-2757: logging crashes on reconfiguration
The way that LogObjectManager transfers LogObjects to the new manager
when logging is reconfigured is prone to crashes since the previous
object manager can be used after the objects have been transferred
out of it. The solution to this is to persist the objects in the
manager for the whole of the manager's lifetime. This require adding
reference counting so that multiple managers can safely share
ownership of the same objects.

Added a TSQA test to exercise this code path. Plugged a LogFormat
leak when creating new TextLogObjects.