TS-2744: remove TSNetAcceptNamedProtocol assertion for unknown protocols
TSNetAcceptNamedProtocol is supposed to accept arbitrary strings
because it lets you implement your own protocol plugin over TLS,
using NPN or ALPN to demux the session. TS-2431 added an assertion
to SSLNextProtocolSet::NextProtocolEndpoint::NextProtocolEndpoint
when an unknown protocol is registered. There's no "unknown" value
in TSProtoType, but it is safe to mark unrecognized protocols as
TS_PROTO_TLS.