TS-2797: Build all manual pages in the doc/reference/api directory
Not all of the manpages are getting built because they are not part
of the man_pages list in doc/conf.py

This patch adds all of the files in the doc/reference/api directory
to the list of manual pages. It also adds the manual page descriptions
to the HTML output.

This closes #82