Document proxy.config.stack_dump_enabled
While debugging the issues we've had with ATS, we needed to generate
core files, but by default ATS handles the SIGSEGV signal and prints
the backtrace to traffic.out. We've find out, though, that there's
a config setting that disables this behavior, allowing core files
to be generated, and decided we should add this information to the
doc as others, specially newcomers to ATS, may find it useful.

This closes #109.