TS-3120: overlapping remap rank when using .include directives
As described in the docs, we use line number for remap rule rank:

  Once these rules are executed we pick the lowest line number as
  the match (which replicates first-match-wins).

However, when we use .include directives to include some other remap
config files, there will be overlapping and conflict with the line
numbers in each other file.

*Examples*

  remap.config
    .include remap1.config
    .include remap2.config

  remap1.config
    map /foo/ https://www.yahoo.com

  remap2.config
    map /foo/bar1 https://www.yahoo.com
    map /foo/bar2 https://www.yahoo.com

When parsing remap1.config, first entry in remap1.config is inserted
with rank 0, second with rank 1. Then parsing remap2.config, the
single entry is inserted with rank 0 again. So the entry in
remap2.config is overlapped with first entry in remap1.config and
takes precedence with second entry. This would confuse customers.

I use count in UrlRewrite::_addToStore for rank. It is used to count
the number of each type of map rules(map, reverse_map,
map_with_referer...). When we insert a new url_mapping, it is stored
in a separate MappingsStore. So there would be no conflict with
each type's count.