TS-3119: add SO_LINGER socket option support
The timeout should be 0 if we turn on SO_LINGER option as we are
using non-blocking socket. I have tried a non-zero timeout, but the
setsockopt failed.

If we add a new sock option
  static unit32_t const SOCK_OPT_LINGER_ON = 4;

Then we could configure
  proxy.config.net.sock_option_flag_in
  proxy.config.net.sock_option_flag_out

to make client connection or server connection LINGER for 0 ms if
needed.