TS-3301: improved TLS ticket rotation support
We all know that it is bad security practice to use the same
password/key all the time. This project tries to rotate TLS session
ticket keys periodically. When an admin runs "traffic_line -x" after
a new ticket key is put in the key file ssl_ticket.key, an event
will be generated and ATS will reconfigure SSL. The keys are read
in all at the same time and the first entry is the most recent key.
A new key is assumed to be put at the beginning of ssl_ticket.key
file and an old key is chopped off at the end from the file.