TS-3403: stop parsing command-line options at the first non-option
ink_args was skipping over file arguments to parse subsequent option
flags. While this allows options and file arguments to intermingle,
it makes it impossible to pass unparsed options down to subcommands.

To solve this, we stop parsing the command line at the first
non-option and declare that the rest of the options are file arguments.