TS-3607: Integrate ats_pagespeed into the automake build
This commit adds support to build the ats_pagespeed plugin directly from
the automake build. It will include the PSOL libraries from a configure
option. It also renames its implementation of the gzip plugin to
ats_pagespeed_gzip to allow coexistence with the upstream version in the
same build.

It also fixes some nits to comply with clang-format.