TS-2150: Transaction milestone log tag
It includes a refactoring for Milestone class attributes to be
requested using public accessor on a private array indexed with an
enum.

Also a minor feature fix, which is aggregate log tag detection, that
matched aggregate string (like SUM FIRST etc..) using strstr() inside
the log tag preventing use of these string to be used in container
field.