Make HostDBRoundRobin::select_best_http take last_failure time into consideration for all RR types
In the current setup it only checks that status of the reals if you use "default RR" (which is actually consistent hashing... but we'll let that slide). This patch consolidates the alive() check into the HostDBInfo struct, and then calls if from all 3 LB mechanisms. Since you can control if/when a host is marked as down in ATS there is no reason to not check.

Issue: TS-3724