TS-3820 Change the default for proxy.config.http.redirect_host_no_port
This will now enable this feature by default, and therefore, should
also be backported to 6.0.0. The point of this option is to avoid
adding the port to the redirected host, unless required. This is
already standard ATS practices, and we made this option to be
backwards compatible. I think for 7.0.0, we should simply remove
this option completely unless someone is using it still at that
point.