TS-4099: add Lua bindings for custom metrics
Implement custom metrics in terms of Lua bindings. We export a
minimal API that constructs metrics and implicitly registers an
evaluation function. We periodically evaluate all the evaluation
functions and update their corresponding metrics.

Metrics are implicitly registered when they are loaded, so there
is no need to mention anything in mgmt/RecordsConfig.cc.

The metrics registration and evaluation syntax looks like this:

  integer 'proxy.node.http.user_agents_total_documents_served' [[
      proxy.process.http.incoming_requests
  ]]