TS-4054 Incorrect ink_assert behavior in Diags.cc
`ink_assert(diags_log == NULL)` is incorrect since there are times
when `blf` correctly has the value NULL. In that case, we would
like for the function to do nothing and return early versus crash
ATS via an assert. That is to say, semantically an early return is
more clear than an assert.