TS-4047: consistently handle multiple RPATH options
Using TS_ADDTO to add the custom locations of libraries results in
incorrect "-rpath" directive to the linker. This is because TS_ADDTO
creates a unique list of tokens that it adds to its target.

e.g. ./configure --with-openssl=/usr/local/openssl --with-zlib=/usr/local/zlib

The above results in LIBTOOL_LINK_FLAGS:
  -rpath /usr/local/openssl/lib /usr/local/zlib/lib

The absence of "-rpath" in front of the second path causes an error
in the linker. This is fixed by introducing a new macro TS_ADD_RPATH_TO
that takes a path and adds it with the "-R " prefix.

Also added the influential environmental variable RPATH that can
be used to give an additional value for the rpath.

This closes #324.