TS-4057 Missing check for system call error
EPERM and EACCES are often consfused for each other.

EPERM means "operation not permitted", ie doing this would not be
in your best interest.
EACCES means "permission denied", meaning if you run this command
as sudo it will work.

Opening a log file will _probably_ never be bad, and even if the
system says so, it shouldn't be bad. So in this case it doesn't
hurt to leave both conditions in here.