TS-4373: TSSslServerContextCreate and TSSslContextDestroy.
TSSslServerContextCreate returns a new SSL Context that's configured
according to the settings in records.config.

This is useful if an extension wants to use the TS_SSL_CERT_HOOK to
control loading of SNI certificates, and still want to respect the
cipher suite and related SSL settings.

Add TSSslContextDestroy method.