TS-4107: proxy.process.ssl.total_success_handshake_count_in has wrong record type
proxy.process.ssl.total_success_handshake_count_in should be the
metrics generated in the core SSL code, and the old compatible name
proxy.process.ssl.total_success_handshake_count should be generated
as a custom metric. The custom metric needs to be registered in
RecordsConfig.cc but it should be of type RECT_PROCESS not RECT_NODE.

This closes #409.