TS-4261: Split stats API from process API.
This is useful for TS-4260 as noted in the related bugs. As part
of that work I set up some statistics to track the performance of
the event loop. Without this change doing that requires bringing
all of the process management support in to the event loop component
which is problematic. It seemed much simpler and better overall to
just split those unrelated items apart.

This closes #516.