TS-4207 Adds better checks to avoid HostDB crashes in 6.x
As much as I dislike this bandaid fix, it does prevent the crashes that we've
been experiencing since upgrading to 6.1.x. It's also in line with other
checks for the validity of the rr done in other places. Since the crash is
rather infrequent (once or twice a day on ~30 machines), it's difficult to
track down the exact commit that broke this.

But I did run a build that reverted all HostDB changes since 5.3.x and that
also fixes the crasher. So it's definitely either a fix triggering an old bug,
or a regression.