TS-4272: Clear memory for new HostDBInfo items
std::map is creating an entry, and we are setting a subset of values. Since we were not clearing the memory that we got, we end up with some non-initialized structure fields-- meaning the HostDBInfo object we return is corrupt.

Fixes #523

(cherry picked from commit edf9ff13db4e63ee9b5003b603439bf2aa6ff19b)