Fix for TS-4276
dnsEvent assumes that lookup_done must return a value, but in error conditions it is not. Since we already have the DNS resolution, it makes more sense to return the non-persistant struct we have instead of returning NULL. This way we can maintain our "contract" of always returning something even if hostdb is full.

Fixes #526