TS-4328 Not retry if any bytes were sent
Before we would still retry if the origin didn't error-- but just took a really long time. So, instead of only checking if the body was sent-- we can check the SM to see if we wrote any header bytes to the wire. If any bytes were sent the request cannot be retried.

This closes #554

This also deprecates the `request_body_start` field, and as such removes it