TS-4328: removing flawed test case
This test case was intended to test the case where the TCP session was established but no bytes were sent. The issue is that user space has no control over the SYN-ACK, accept() simply is user-space saying "i want this connection". So because of that this test is useless, and can be removed

(cherry picked from commit 770ca6fb2f10ea91d66a91ebca7e7689be174cb8)