TS-4366 Uninitialized stack value used in mp4 plugin
It is possible that there are cases where IOBufferReaderCopy() does not modify
the input buffer (copy 0 bytes) which leaves the buffer uninitialized which is
undesirable since the buffers are always allocated on the stack.

Addressed it in one of 2 ways:
(1) memset(buffer, 0, sizeof(buffer)) or
(2) check IOBufferReaderCopy() return value and handle accordingly.

These changes are meant to only address using uninitialized values allocated on
the stack, avoiding bigger changes since regression tests are not available to
properly verify functionality.

This closes #656