TS-4375: Fix PCRE link issues on Darwin.
Darwin has a copy of PCRE in the SDK, but it has JIT disabled. When
we use pkg-config to find the PCRE version, we find the Homebrew
version in /usr/local which does have JIT enabled. Now, we end up
searching for Expat in the SDK so the SDK ends up in our library
search path causing us to find the wrong version of libpcre. Link
failures ensue.

Since it is so trivial to just use expat from Homebrew, remove the
SDK path search and tidy up the XML m4 macros to be a bit more
informative.