TS-4388: Add an API test for TSHttpTxnParentProxySet.
Add an API test for TSHttpTxnParentProxySet. This test sends a
request to the API test synserver, and expects a response. The quirk
is that it can't get the right response unless TSHttpTxnParentProxySet()
is called to parent proxy to the syn server.

This test currently crashes, so it is only enabled at REGRESSION_TEST_FATAL
(level 4).