TS-4020 Use cache-key URL for some plugins
cache_promote and background_fetch both use the pristine URL
as the hash for various internal lookup features. This works
well, except when it doesn't. For example, when mixed with
using the cachekey.so plugin.

This changes the behavior to use the cachekey URL. This is not
awesomely efficient at this point, but a future fix will use a
new API that allows introspection into the actual cache key used
by the ATS core instead.