TS-4395 Add new plugin, remap_purge
This is similar to the existing plugin to purge based on a genID stored in a
persistent storage. The difference is that the purging is done exclusively via
a restful API, and has little (no) overhead on performance (since the
generation ID is always in memory).

Example:

map http://example.com/p1  http://p1.example.com
  @plugin=remap_purge.so @pparam=--path=__secret__  @pparam=--state=example_p1

And to purge:

  $ curl -s -D - -X PURGE http://example.com/p1/__secret__
  HTTP/1.1 200 OK
  Date: Sat, 30 Apr 2016 00:09:34 GMT
  Connection: close
  Server: ATS/7.0.0
  Content-Length: 39
  Content-Type: text/html

  PURGED http://example.com/p1