TS-4436: Move hosts file implementation to `do_dns`
This moves the hosts file overrides down to the DNS layer instead of the HostDBInfo layer. This means that each port will get its own entry in hostdb, which will exist as its own HostDBInfo just like all the rest of the entries (with a TTL of the time remaining until the next hostdb sync). This also means we only have to do the check in the map on DNS lookup instead of on each probe().

In addition to fixing the down status issues, this also means we no longer need to keep old copies of the strings etc. since they are copied once lookup_done is called.