TS-4441: Stop dumping core on fatal errors.
On a fatal error, the core file is rarely useful, and there is a
real risk of filling up a disk with cores when we get a bad
configuration. Rather than dumping core when logging a fatal error,
just exit.

Add a new API ink_abort() that logs a fatal error followed by
abort(3) rather than exit(3). Use this in ink_assert() and other
places where it makes sense to dump core.