TS-3959: requests are retryable only if bytes were sent and there was no connection error
The root issue that TS-4328 and TS-3440 were trying to fix was that ATS was too aggressive in retries to origin (namely after the origin had already recieved the request). The previous patches attempted to have ATS not retry assuming bytes were sent on the wire-- sadly TS-4328 actually only checks that we set bytes to be written to the wire. To fix this we simply check if there was a connection failure to origin, if so we'll assume that it is retryable (since connection failures to origin should only be due to errors writing/connecting etc.).

(cherry picked from commit 736009b973ad44ef4c7185f5f3bfb778903e4316)