Fix ParentProxySet configuration race.
The first part of the problem occurs because the HTTP state machine
samples the parent proxy configuration at the start of a transaction.
We have to delay the start of the test until we know the configuration
has been applied.

The second pars occurs because the regression driver starts the
next exclusive test as soon as the current one has set its status.
This means that we need to tear down the synserver and any other
exclusive state before setting the test status.

(cherry picked from commit 55c57b39ce0bf51ec2b10fc97e401c2b6588f1ae)