TS-3245: Allow multiple plugins to safely use getopt(3).
Allow multiple global plugins to co-exist in the plugin.config
file. Also, made the same changes to allow remap plugins
to co-exist with each other in the remap.config file.

Specifically, on freebsd and darwin platforms we use 'optreset =
1' to reset the getopt_long() command line parser.  On Linux/glibc
we use 'optind = 0' to reset the parser.  The parser must be reset
between uses such as first loading one plugin and then another.

The setting of these variables is now performed in ATS core before
the plugin initialization functions are called. The setting of these
variables inside of the individual plugins has been removed. If a
plugin needs to call getopt_long() outside of the plugin global or
remap intializations, it should use the code in proxy/Plugin.cc as
an example.

This closes #696.