TS-4570: Return to TS_HTTP_OS_DNS_HOOK if server address is set.
Using HostDB, Traffic Server will try to connect to all the resolved
IP addresses if the connection fail. However if a plugin sets the
address using some out-of-band knowledge and TSHttpTxnServerAddrSet,
it onlt gets to set the address once.

This change returns the state machine to the OS_DNS_HOOK API callout
so that the plugin can have an opportunity to set a different server
address.