TS-4612: Proposal: InactivityCop Optimize
Load the NetVCs from open_list to cop_list before next InactivityCop
runs. NetHandler will remove NetVC from cop_list if it is triggered. As
the NetHandler runs, the number of NetVCs in the cop_list is decreasing.
Lots of NetVCs will be removed from cop_list because the NetHandler runs
100 times maximum between InactivityCop runs. Therefore we don't have to
check all the NetVCs as much as open_list.