TS-4472 Fixes http_current_active_client_connections_stat metric
The issue here is that there are two m_active member variables, in
both Http1ClientSession and Http1ClientTransaction. However, the former
contains the latter, so this patch eliminates the m_active from the
session class, and then consistently sets and checks the m_active from
the http1ClientTransaction class.