TS-4674: Remove useless assert statement
Now that we have clean allocations (instead of clobbering existing things) there is no need to have this assertion. In practice this assertion is actually incorrect, because in the case where we want to extend the lifetime of a stale record (since the response we got was broken) we'll fail this assert.

This closes #809