TS-4707 : Parent Consistent Hash Selection - Enhancements
Enable the parent consistent hash selection feature to hash an
arbitrary string instead of the URL stored in the request header.
The string is read from the request header from a custom internal
MIME field '@ATS_PARENT_SELECTION_URL'. This arbitrary string may
be set directly by a plugin.

The following API calls may also be used in which case the string
must be a valid URL rather than arbitrary:
 tsapi TSReturnCode TSHttpTxnParentSelectionUrlSet();
 tsapi TSReturnCode TSHttpTxnParentSelectionUrlGet();
 tslua ts.http.set_parent_selection_url()
 tslua ts.http.get_parent_selection_url()