TS-4785: Stale while revalidate plugin memory corruption.
This fixes a couple of memory corruption bugs in the stale_while_revalidate
plugin.

First, when copying the client sockaddr, we need to copy the specific
address family structure, not the untyped struct sockaddr.

Next, we zero when allocating to make initialization more reliable.

Finally, we deal with receiving TS_EVENT_HTTP_CACHE_LOOKUP_COMPLETE
multiple times, which otherwise leads to a double free.