TS-4825: Add the TSVConnInternalSet() API.
If you have a protocol plugin that generates HTTP requests, you
typically want those to be treated as external client requests not
as internal requests.

The TSVConnInternalSet() allows a plugin to mark a TSVConn as either
internal or external. It should be called immediately after the
TSVConn is created in order to avoid any race conditions with
any consumers.