TS-4840: Ensure unallocated stats cannot be reattached.
If traffic_server is restarted, any plugin stats that have been
created are still registered in traffic_manager so we can find them
on startup with TSStatFindName. However, because traffic_server is
a new process, we don't have them allocated in the plugin RSB so
things go awry from there.

The solution is to not allow TSStatFindName to succeed unless there
is a also plugin TSB entry allocated for this stat.