TS-4870: Avoid marking storage offline multiple times
Added a CacheDisk:online flag because we cannot relay on DISK_BAD(disk)
macro to identify if we already marked it bad. The problem is that in the
common use-case in handle_disk_failure(), the disk is already bad when
mark_storage_offline() is called, so we can't depend on the good->bad state
transition to know when to update the accounting.

(cherry picked from commit 7421eabf0f74da0005c25cc199b79f28ff5159de)