TS-4875: Hoist SSL index creation into library initialization.
The SSL index used for server certificate validation can only be
created once, so we need to do it in the library initialization,
not the client context initialization.

Following the SSL utilities conventions, wrap a helper API around
the raw SSL data index accesses.

Replace calls to the obsolete OpenSSL SSL_set_app_data and
SSL_get_app_data APIs with the new utilities. We don't need to store
the SSLNetVConnection back pointer is 2 separate indices.