TS-4882: Implement TCP Fast Open for server sockets.
Add TCP Fast Open support to the proxy.config.net.sock_option_flag_out
configuration option.

Implement detection of TFO support in SocketManager and use that
to guarantee that we don't try to use TFO on hosts where is is not
supported or enabled.

Add TFO support to both SSLNetVConnection and UnixNetVConnection,
enabling it if the corresponding NetVCOption is set. We only enable
TFO on server sessions that are going to do TLS or are handling
idempotent HTTP methods.