TS-4903: EISCONN errors when using TCP Fast Open.
The write VIO is reset when we reuse a VC, so we need to explicitly
track whether we have connected a TCP Fast Open socket. In the
previous patch we eschewed using the is_connected state of the
Connection, but that is explicitly what we want so we should just
use it and allow the Connection to be not connected until the Fast
Open connects it.