TS-4399: Management API messes up proxy options.
TS-4399: Management API breaks diagnostic log rotation
TS-4400: TSProxyStateSet persist cache clearing across restart

The two issues above are related in that they both deal with the
management API not correctly handling proxy flags.

For TS-4399, it was because the management API was not aware
of traffic_manager setting extra proxy options. This was fixed
by providing CoreAPI a callback to get extra proxy options from
traffic_manager.

For TS-4400, it was because the management API was not properly
clearing optional flags between proxy reboots. This was fixed
by resetting the proxy options before each reboot.

This patch centralizes where traffic_server starts to inside
CoreAPI's ProxyStateSet. This is good because we reduce the
number of places we deal with traffic_server options. Everything
is simply handled by the proxy_options_callback.

We use proxy_options as persistent proxy options storage.  Then
have lmgmt->startProxy() take in an argument for one-time flags.

This closes #1073.