TS-4925: Manager pollMgmtProcessServer stuck with EBADF.
If the select in pollMgmtProcessServer fails with an actual error
(EBADF, for example), it just returned back into the select loop
and was unable to recover. This might happen if the server file
descriptor is -1, or otherwise stale.

The fix is to not put -1 into the select set, and to break the loop
if select returns an error.