TS-4952: Improve abort messages for mutex failures.
Although the ink_mutex APIs claimed to return error values, they
actually abort on failure. Improve the abort to format a message
that includes the error, and don't bother to return error values
any more since these can't fail.

Since we now check for mutex manipulation errors, error checking
mutexes (the default type on FreeBSD) will fail when we unlock a
Thread mutex from a different thread than the one that created it.
To solve this, we introduce TSThreadWait so that the creating thread
can safely spawn a thread, wait for it and then destroy it.