TS-4955: Allow Plugin VC accept continuation to be unlocked.
We never need to hold the continuation lock when delivering
NET_EVENT_ACCEPT, so allow a NULL continuation mutex in PluginVC to
prevent lock contention on the plugin_http_accept object.

We still hold the lock if it is present since in the intercept cases,
the continuation can come from plugins that may be expecting us to
lock. This lock is now optional, however, and it is not an error for
the plugin to omit the mutex.