TS-4947: Fix log collation hosts configuration.
In the Lua logging configuration, we were failing to correctly
configure log collation hosts. The main fix is to simply return
true when we actually succeeded.

Fixed a bug in lua_scoped_stack where we failed to count the number
of values we pushed, which messes up the stack on return.

Updated the LogHost display to show failover hosts.

Added some tests for the supported syntaxes.