Fix CoreAPI sleeping deadlock
This change was made because the ProxyStateSet() call was sleeping
in the same thread as traffic_manager was running, so the
condition in the do/while loop could never be updated by traffic_manager.

The solution is to differentiate API calls and local calls to
CoreAPI by using the coreapi_sleep flag.