TS-5040: Forward CONNECT without parent proxying.
Unless parent proxy is enabled, CONNECT methods always result in
setting up a direct tunnel to the origin server. Parent proxy,
however, has the option of forwarding the CONNECT to the next hop
and allowing it to set up the tunnel.  This change adds the
proxy.config.http.forward_connect_method configuration parameter
that can be enabled to forward CONNECT methods even if we are not
parent proxying.