Fix chmod permission errors & code refactor
This was needed because traffic.out, unlike squid.blog, is created
before ATS drops superuser status. Thus, we cannot use a traditional
chmod(2) on traffic.out since traffic.out is owned by root when
we want to change it as nobody.

The trick here is to use the CAP_FOWNER capability to bypass permission
checks.